+++ bin/report.dart
@@ -4,16 +4,14 @@ import 'common.dart';
 
 const reportCycles = 10; // for upper layers
 
-printReport({required List<Hop> stat, required int hops, String? target, bool last = false, int indent = 4}) {
+printReport(List<Hop> stat, int hops) {
   if (hops > 0) {
-    if (target != null) print('Ping $target');
-    print(sprintf('%*s%-*s %s', [indent, '', dnsEnable ? maxHostname : maxHostaddr, hostTitle, statTitle]));
+    print(sprintf('%*s%-*s %s', [lindent, '', dnsEnable ? maxHostname : maxHostaddr, hostTitle, statTitle]));
     for (int i = 0; i < hops; i++) {
       String no = sprintf('%2d. ', [i + 1]);
-      print(sprintf('%*s%s', [indent, no, (stat[i].data.sent > 0) ? stat[i] : '']));
-      for (int j = 1; j < stat[i].addr.length; j++) { print(sprintf('%*s%s', [indent, '', stat[i].lpart(j)])); }
+      print(sprintf('%*s%s', [lindent, no, (stat[i].data.sent > 0) ? stat[i] : '']));
+      for (int j = 1; j < stat[i].addr.length; j++) { print(sprintf('%*s%s', [lindent, '', stat[i].lpart(j)])); }
     }
   }
-  if (!last) print('');
 }